Pushing up changes from a year ago... inconsistent state..
It's been a while since my last commit. I think I've attempted to refactor
this module and kind of left it a state where some things are not finished.
This commit is just to make sure I copy these changes into github and figure out
how to fix things later